Fix multiple inclusion errors for ethernet structure
Inclusion for the low-level ethernet headers in the rp-pppoe plugin is
quite a mess and does not work for musl libc as-is.  Since this code is
Linux-only and uses a static config.h, the patch gets rid of the
pointless flexibility, replacing it by a single inclusion of
linux/if_pppox.h in pppoe.h.