Restore effectiveness of ipcp-max-failures option
In release 2.4.5, pppd increased the maxnakloops value from 5 to 100
to cope with the behavior of certain mobile networks.

Before that change, the maxnakloops value could be configured through
the ipcp-max-failures option, however the change hardcoded the higher
value into the middle of initialization code, rendering the option
ineffective.

This commit removes the hardcoded override in favor of changing the
default value for the option, retaining the new behavior (although 100
seems a bit excessive), while also making the option effective again.